For #9728: Add actionbar search for language list.
